+++ src/Particle/Kernel/Size.cl
@@ -2,7 +2,7 @@
 
 void kernel sizeUpdate(__global ParticleData *data,
         __global ModuleParamSizeOverLifetime *moduleParams) {
-    size_t id = get_global_id(0);
+    uint id = get_global_id(0);
     __global ParticleData *particle = &data[id];
 
     particle->size = moduleParams->size.min + moduleParams->size.max * (particle->age / particle->lifeTime);
